+++ app/Http/Controllers/SubjectController.php
@@ -24,6 +24,9 @@ class SubjectController extends Controller
      */
     public function create()
     {
+        if(auth()->user()->cannot('viewAny', Subject::class)){
+            abort(403);
+        }
         $courses = Course::all();
         return view('components.subjects.create', compact('courses'));
     }
@@ -33,11 +36,15 @@ class SubjectController extends Controller
      */
     public function store(SubjectRequest $request)
     {
-        try {
-            Subject::create($request->validated());
-            return redirect()->route('subjects.index')->with('success','Subject created successfully!');
-        } catch (\Exception $e) {
-            return redirect()->route('subjects.index')->with('error','Something went wrong during subject creation! ' .$e->getMessage());
+        if($request->user()->cannot('create', Subject::class)){
+            abort(403);
+        }else{
+            try {
+                Subject::create($request->validated());
+                return redirect()->route('subjects.index')->with('success','Subject created successfully!');
+            } catch (\Exception $e) {
+                return redirect()->route('subjects.index')->with('error','Something went wrong during subject creation! ' .$e->getMessage());
+            }
         }
     }
 
@@ -54,6 +61,9 @@ class SubjectController extends Controller
      */
     public function edit(Subject $subject)
     {
+        if(auth()->user()->cannot('update',$subject)){
+            abort(403);
+        }
         $courses = Course::all();
         return view('components.subjects.edit', compact('subject','courses'));
     }
@@ -63,6 +73,10 @@ class SubjectController extends Controller
      */
     public function update(SubjectRequest $request, Subject $subject)
     {
+        if($request->user()->cannot('update',$subject)){
+            abort(403);
+        }
+
         try {
             $subject->update($request->validated());
             return redirect()->route('subjects.index')->with('success','Subject updated successfully!');
@@ -76,6 +90,9 @@ class SubjectController extends Controller
      */
     public function destroy(Subject $subject)
     {
+        if(auth()->user()->cannot('delete',$subject)){
+            abort(403);
+        }
         try {
             $subject->delete();
             return redirect()->route('subjects.index')->with('success','Subject updated successfully!');
